requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
bardapi==0.1.34
beautifulsoup4==4.12.2
blis==0.7.11
browser-cookie3==0.19.1
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.5
colorama==0.4.6
comm==0.1.4
confection==0.1.3
contourpy==1.1.1
cssutils==2.7.1
cycler==0.11.0
cymem==2.0.8
dataclasses-json==0.6.0
debugpy==1.8.0
decorator==5.1.1
deep-translator==1.11.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.4
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.2
google-api-core==2.11.1
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.0
huggingface-hub==0.17.3
hyperframe==6.0.1
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
langchain==0.0.300
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.0.40
lit==17.0.1
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
numexpr==2.8.6
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openai==0.28.1
OpenAIAuth==3.0.0
packaging==23.1
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip-chill==1.0.3
platformdirs==3.10.0
preshed==3.0.9
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.19.0
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
revChatGPT==6.8.6
rich==13.5.3
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.21
srsly==2.4.8
stack-data==0.6.2
sympy==1.12
tenacity==8.2.3
thinc==8.1.12
threadpoolctl==3.2.0
tiktoken==0.5.1
tls-client==0.2.2
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
transformers==4.33.3
triton==2.0.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.16
wasabi==1.1.2
wcwidth==0.2.6
yarl==1.9.2
youtube-transcript-api==0.6.1
